requirements.txt
Flask
MarkupSafe
Pillow
flask_bootstrap
